feature: sorcerer now supports rule return type action code which includes the C/C++ RAII initial assignment expressions, e.g.
```
format_id > [SORAST *Format = NULL]:
   ....
```

as used in the f77 example code: the result is that the generated code now does not cause MSVC to complain any more about "possibly uninitialized variable" issues in the generated code. This is not bogus, as this also helps us as sorcerer grammar writers: the initial value expression helps to cover for when we forgot to set that variable in any/each of the rule branches' action code.

---

TODO: something to consider for pccts/antlr itself as well: that one has its own set of codegen/dump functions, so this patch must be transported to that codebase for this to work there.